chore: refactor trigger code to multiple files
